waypointman + safe_mode: Fix Gui vs model parts
Waypointman: Move GLrelated stuff from waypoint_man to
waypointman_gui (3533)

safe_mode: Split into gui and model parts.
