+++ CMakeLists.txt
@@ -916,6 +916,7 @@ set(
   include/rest_server.h
   include/route.h
   include/routeman.h
+  include/safe_mode.h
   include/select.h
   include/select_item.h
   include/semantic_vers.h
@@ -998,7 +999,8 @@ set(
   include/S57ObjectDesc.h
   include/S57QueryDialog.h
   include/S57Sector.h
-  include/safe_mode.h
+  include/safe_mode_gui.h
+  include/select_item.h
   include/SendToGpsDlg.h
   include/SendToPeerDlg.h
   include/shaders.h
@@ -1095,6 +1097,7 @@ set(MODEL_SRC
   ${CMAKE_SOURCE_DIR}/src/s57featuredefns.cpp
   ${CMAKE_SOURCE_DIR}/src/s57reader.cpp
   ${CMAKE_SOURCE_DIR}/src/s57RegistrarMgr.cpp
+  ${CMAKE_SOURCE_DIR}/src/safe_mode.cpp
   ${CMAKE_SOURCE_DIR}/src/select.cpp
   ${CMAKE_SOURCE_DIR}/src/select_item.cpp
   ${CMAKE_SOURCE_DIR}/src/semantic_vers.cpp
@@ -1185,7 +1188,7 @@ set(
   ${CMAKE_SOURCE_DIR}/src/s57chart.cpp
   ${CMAKE_SOURCE_DIR}/src/s57obj.cpp
   ${CMAKE_SOURCE_DIR}/src/S57QueryDialog.cpp
-  ${CMAKE_SOURCE_DIR}/src/safe_mode.cpp
+  ${CMAKE_SOURCE_DIR}/src/safe_mode_gui.cpp
   ${CMAKE_SOURCE_DIR}/src/SencManager.cpp
   ${CMAKE_SOURCE_DIR}/src/SendToGpsDlg.cpp
   ${CMAKE_SOURCE_DIR}/src/SendToPeerDlg.cpp
