+++ include/safe_mode_gui.h
@@ -1,0 +1,21 @@
+/**
+ * Safe mode GUI handling.
+ *
+ * Administrate the safe mode state which is true if openpcn should run in
+ * safe mode without OpenGL, plugins, etc.
+ */
+
+#include <wx/window.h>
+
+#include "OCPNPlatform.h"
+#include "safe_mode.h"
+
+namespace safe_mode {
+
+/**
+ * Check if the last start failed, possibly invoke user dialog and set
+ * safe mode state.
+ */
+void check_last_start();
+
+}
