+++ src/safe_mode.cpp
@@ -8,7 +8,6 @@
 
 #include "base_platform.h"
 #include "cmdline.h"
-#include "gui_lib.h"
 #include "ocpn_utils.h"
 
 #include "safe_mode.h"
@@ -17,22 +16,16 @@ extern BasePlatform* g_BasePlatform;
 
 namespace safe_mode {
 
-static const char* LAST_RUN_ERROR_MSG =
-    _("<p>The last opencpn run seems to have failed. Do you want to run\n"
-      "in safe mode without plugins and other possibly problematic\n"
-      "features?\n</p><br/></br><p>You may consider visiting the <a href=\"https://github.com/OpenCPN/OpenCPN/wiki/OpenCPN-5.8-known-issues\">list of known issues</a>.</p>");
-
 #ifdef _WIN32
 static std::string SEP("\\");
 #else
 static std::string SEP("/");
 #endif
 
-static const int TIMEOUT_SECONDS = 15;
 
-static bool safe_mode = false;
+bool safe_mode = false;
 
-static std::string check_file_path() {
+std::string check_file_path() {
   std::string path = g_BasePlatform->GetPrivateDataDir().ToStdString();
   path += SEP;
   path += "startcheck.dat";
@@ -46,26 +39,6 @@ void set_mode(bool mode) {
 
 bool get_mode() { return safe_mode; }
 
-/**
- * Check if the last start failed, possibly invoke user dialog and set
- * safe mode state.
- */
-void check_last_start() {
-  std::string path = check_file_path();
-  if (!ocpn::exists(path)) {
-    std::ofstream dest(path, std::ios::binary);
-    dest << "Internal opencpn use" << std::endl;
-    dest.close();
-    return;
-  }
-  long style = wxYES | wxNO | wxNO_DEFAULT | wxICON_QUESTION;
-  auto dlg = new OCPN_TimedHTMLMessageDialog(0, LAST_RUN_ERROR_MSG,
-                                             _("Safe restart"), TIMEOUT_SECONDS,
-                                             style, false, wxDefaultPosition);
-  int reply = dlg->ShowModal();
-  safe_mode = reply == wxID_YES;
-  dlg->Destroy();
-}
 
 /** Mark last run as successful. */
 void clear_check() { remove(check_file_path().c_str()); }
